requirements.txt
flask-appconfig>=0.10
flask-bootstrap
flask-nav
flask-debug
flask-wtf
flask-babel
flask-sqlalchemy
flask-script
flask-migrate